@@ -11,18 +11,18 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^7.2.14",
+    "@angular/animations": "^8.2.2",
     "@angular/cdk": "^7.3.7",
-    "@angular/common": "^7.2.14",
-    "@angular/compiler": "^7.2.14",
-    "@angular/core": "^7.2.14",
-    "@angular/forms": "^7.2.14",
+    "@angular/common": "^8.2.2",
+    "@angular/compiler": "^8.2.2",
+    "@angular/core": "^8.2.2",
+    "@angular/forms": "^8.2.2",
     "@angular/http": "^7.2.14",
     "@angular/material": "^7.3.7",
-    "@angular/platform-browser": "^7.2.14",
-    "@angular/platform-browser-dynamic": "^7.2.14",
-    "@angular/platform-server": "^7.2.14",
-    "@angular/router": "^7.2.14",
+    "@angular/platform-browser": "^8.2.2",
+    "@angular/platform-browser-dynamic": "^8.2.2",
+    "@angular/platform-server": "^8.2.2",
+    "@angular/router": "^8.2.2",
     "@iconscout/unicons": "^1.0.1",
     "body-parser": "^1.19.0",
     "core-js": "^2.6.5",
@@ -31,7 +31,7 @@
     "install": "^0.12.2",
     "mongoose": "^5.5.5",
     "morgan": "^1.9.1",
-    "rxjs": "^6.5.1",
+    "rxjs": "^6.5.2",
     "rxjs-compat": "^6.5.1",
     "serve-favicon": "^2.5.0",
     "zone.js": "^0.9.1"
@@ -39,8 +39,8 @@
   "devDependencies": {
     "@angular-devkit/build-angular": "^0.13.8",
     "@angular/cli": "^8.2.2",
-    "@angular/compiler-cli": "^7.2.14",
-    "@angular/language-service": "^7.2.14",
+    "@angular/compiler-cli": "^8.2.2",
+    "@angular/language-service": "^8.2.2",
     "@types/jasmine": "^3.3.12",
     "@types/jasminewd2": "~2.0.6",
     "@types/node": "^12.0.0",
@@ -57,6 +57,6 @@
     "rxjs-tslint": "^0.1.7",
     "ts-node": "~8.1.0",
     "tslint": "^5.16.0",
-    "typescript": "^3.2.4"
+    "typescript": "^3.5.3"
   }
 }
